+++ R/06-LoadPars.R
@@ -83,14 +83,34 @@ LoadPars <- function(itemtype, itemloc, lambdas, zetas, guess, upper, fulldata, 
                                 paste('ak', 0L:(K[i]-1L), sep=''),
                                 paste('d', 0L:(K[i]-1L), sep=''))
             }
-        } else if(any(itemtype[i] == c('2PL', '3PL', '3PLu', '4PL'))){
+        } else if(any(itemtype[i] == c('2PL', '3PL', '3PLu', '4PL', '5PL'))){
             if(K[i] != 2L)
                 stop(paste0('Item ', i, ' requires exactly 2 unique categories'), call.=FALSE)
             val <- c(lambdas[i,], zetas[[i]], guess[i], upper[i])
             fp <- c(estLambdas[i, ], TRUE, FALSE, FALSE)
-            if(any(itemtype[i] == c('3PL', '4PL'))) fp[length(fp)-1L] <- TRUE
-            if(any(itemtype[i] == c('3PLu', '4PL'))) fp[length(fp)] <- TRUE
+            if(any(itemtype[i] == c('3PL', '4PL', '5PL'))) fp[length(fp)-1L] <- TRUE
+            if(any(itemtype[i] == c('3PLu', '4PL', '5PL'))) fp[length(fp)] <- TRUE
             names(val) <- c(paste('a', 1L:nfact, sep=''), 'd', 'g','u')
+            if(itemtype[i] == '5PL'){
+                if(nfact != 1L) stop('5PL model currently limited to unidimensional structures',
+                                     call.=FALSE)
+                val <- c(val, logS=0)
+                fp <- c(fp, TRUE)
+            }
+        } else if(itemtype[i] == 'CLL'){
+            if(K[i] != 2L)
+                stop(paste0('Item ', i, ' requires exactly 2 unique categories'), call.=FALSE)
+            if(nfact != 1L) stop('CLL model currently limited to unidimensional structures',
+                                 call.=FALSE)
+            val <- c(b = -zetas[[i]])
+            fp <- TRUE
+        } else if(itemtype[i] == 'ULL'){
+            if(nfact != 1L) stop('ULL model currently limited to unidimensional structures',
+                                 call.=FALSE)
+            val <- c(lambdas[i,], zetas[[i]])
+            fp <- c(estLambdas[i, ], rep(TRUE, K[i]-1L))
+            names(val) <- c(paste('eta', 1L:nfact, sep=''),
+                            paste('log_lambda', 1L:(K[i]-1L), sep=''))
         } else if(any(itemtype[i] == c('2PLNRM', '3PLNRM', '3PLuNRM', '4PLNRM'))){
             val <- c(lambdas[i,], zetas[[i]], guess[i], upper[i],
                      0, rep(.5, K[i] - 2L), rep(0, K[i]-1L))
@@ -306,6 +326,50 @@ LoadPars <- function(itemtype, itemloc, lambdas, zetas, guess, upper, fulldata, 
             next
         }
 
+        if(any(itemtype[i] == '5PL')){
+            pars[[i]] <- new('fivePL',
+                             par=startvalues[[i]],
+                             est=freepars[[i]],
+                             parnames=names(freepars[[i]]),
+                             nfact=nfact,
+                             itemclass=9L,
+                             nfixedeffects=nfixedeffects,
+                             ncat=2L,
+                             any.prior=FALSE,
+                             prior.type=rep(0L, length(startvalues[[i]])),
+                             fixed.design=fixed.design.list[[i]],
+                             lbound=rep(-Inf, length(startvalues[[i]])),
+                             ubound=rep(Inf, length(startvalues[[i]])),
+                             prior_1=rep(NaN,length(startvalues[[i]])),
+                             prior_2=rep(NaN,length(startvalues[[i]])))
+            tmp2 <- parnumber:(parnumber + length(freepars[[i]]) - 1L)
+            pars[[i]]@parnum <- tmp2
+            parnumber <- parnumber + length(freepars[[i]])
+            next
+        }
+
+        if(any(itemtype[i] == 'CLL')){
+            pars[[i]] <- new('cll',
+                             par=startvalues[[i]],
+                             est=freepars[[i]],
+                             parnames=names(freepars[[i]]),
+                             nfact=nfact,
+                             itemclass=9L,
+                             nfixedeffects=nfixedeffects,
+                             ncat=2L,
+                             any.prior=FALSE,
+                             prior.type=rep(0L, length(startvalues[[i]])),
+                             fixed.design=fixed.design.list[[i]],
+                             lbound=rep(-Inf, length(startvalues[[i]])),
+                             ubound=rep(Inf, length(startvalues[[i]])),
+                             prior_1=rep(NaN,length(startvalues[[i]])),
+                             prior_2=rep(NaN,length(startvalues[[i]])))
+            tmp2 <- parnumber:(parnumber + length(freepars[[i]]) - 1L)
+            pars[[i]]@parnum <- tmp2
+            parnumber <- parnumber + length(freepars[[i]])
+            next
+        }
+
         if(any(itemtype[i] == c('2PLNRM', '3PLNRM', '3PLuNRM', '4PLNRM'))){
             pars[[i]] <- new('nestlogit',
                              par=startvalues[[i]],
@@ -358,6 +422,28 @@ LoadPars <- function(itemtype, itemloc, lambdas, zetas, guess, upper, fulldata, 
                              nfact=nfact,
                              ncat=K[i],
                              itemclass=2L,
+                             nfixedeffects=nfixedeffects,
+                             any.prior=FALSE,
+                             prior.type=rep(0L, length(startvalues[[i]])),
+                             fixed.design=fixed.design.list[[i]],
+                             est=freepars[[i]],
+                             lbound=rep(-Inf, length(startvalues[[i]])),
+                             ubound=rep(Inf, length(startvalues[[i]])),
+                             prior_1=rep(NaN,length(startvalues[[i]])),
+                             prior_2=rep(NaN,length(startvalues[[i]])))
+            tmp2 <- parnumber:(parnumber + length(freepars[[i]]) - 1L)
+            pars[[i]]@parnum <- tmp2
+            parnumber <- parnumber + length(freepars[[i]])
+            next
+        }
+
+        if(itemtype[i] == 'ULL'){
+            pars[[i]] <- new('ull',
+                             par=startvalues[[i]],
+                             parnames=names(freepars[[i]]),
+                             nfact=nfact,
+                             ncat=K[i],
+                             itemclass=9L,
                              nfixedeffects=nfixedeffects,
                              any.prior=FALSE,
                              prior.type=rep(0L, length(startvalues[[i]])),