+++ R/mirt.R
@@ -152,10 +152,29 @@
 #'     \deqn{P(x = 1|\theta, \psi) = g + \frac{(u - g)}{
 #'       1 + exp(-(a_1 * \theta_1 + a_2 * \theta_2 + d))}}
 #'   }
+#'   \item{5PL}{
+#'     Currently restricted to unidimensional models
+#'     \deqn{P(x = 1|\theta, \psi) = g + \frac{(u - g)}{
+#'       1 + exp(-(a_1 * \theta_1 + d))^S}}
+#'     where \eqn{S} allows for asymmetry in the response function and
+#'     is transformation constrained to be greater than 0 (i.e., \code{log(S)} is estimated rather than \code{S})
+#'   }
+#'   \item{CLL}{
+#'     Complementary log-log model (see Shim, Bonifay, and Wiedermann, 2022)
+#'     \deqn{P(x = 1|\theta, b) = 1 - exp(-exp(\theta - b))}
+#'     Currently restricted to unidimensional dichotomous data.
+#'   }
 #'   \item{graded}{
 #'     The graded model consists of sequential 2PL models,
 #'     \deqn{P(x = k | \theta, \psi) = P(x \ge k | \theta, \phi) - P(x \ge k + 1 | \theta, \phi)}
 #'     Note that \eqn{P(x \ge 1 | \theta, \phi) = 1} while \eqn{P(x \ge K + 1 | \theta, \phi) = 0}
+#'   }
+#'   \item{ULL}{
+#'     The unipolar log-logistic model (ULL; Lucke, 2015) is defined the same as
+#'     the graded response model, however
+#'     \deqn{P(x \le k | \theta, \psi) = \frac{\lambda_k\theta^\eta}{1 + \lambda_k\theta^\eta}}.
+#'     Internally the \eqn{\lambda} parameters are exponentiated to keep them positive, and should
+#'     therefore the reported estimates should be interpreted in log units
 #'   }
 #'   \item{grsm}{
 #'     A more constrained version of the graded model where graded spacing is equal across item
@@ -307,6 +326,12 @@
 #'     \item \code{'2PL'}, \code{'3PL'}, \code{'3PLu'}, and \code{'4PL'} - 2-4 parameter logistic model,
 #'       where \code{3PL} estimates the lower asymptote only while \code{3PLu} estimates the upper asymptote only
 #'       (Lord and Novick, 1968; Lord, 1980)
+#'     \item \code{'5PL'} - 5 parameter logistic model to estimate asymmetric logistic
+#'      response curves. Currently restricted to unidimensional models
+#'     \item \code{'CLL'} - complementary log-log link model.
+#'       Currently restricted to unidimensional models
+#'     \item \code{'ULL'} - unipolar log-logistic model (Lucke, 2015). Note the use of this itemtype
+#'       will automatically use a log-normal distribution for the latent traits
 #'     \item \code{'graded'} - graded response model (Samejima, 1969)
 #'     \item \code{'grsm'} - graded ratings scale model in the
 #'       classical IRT parameterization (restricted to unidimensional models; Muraki, 1992)
@@ -475,6 +500,8 @@
 #'       \code{quadpts} is increased to 121, and this method is only applicable for
 #'       unidimensional models estimated with the EM algorithm
 #'    }
+#'
+#'    Note that when \code{itemtype = 'ULL'} then a log-normal(0,1) density is used to support the unipolar scaling
 #' @param survey.weights a optional numeric vector of survey weights to apply for each case in the
 #'   data (EM estimation only). If not specified, all cases are weighted equally (the standard IRT
 #'   approach). The sum of the \code{survey.weights} must equal the total sample size for proper
@@ -550,7 +577,9 @@
 #'     \item{NCYCLES}{maximum number of EM or MH-RM cycles; defaults are 500 and 2000}
 #'     \item{MAXQUAD}{maximum number of quadratures, which you can increase if you have more than
 #'       4GB or RAM on your PC; default 20000}
-#'     \item{theta_lim}{range of integration grid for each dimension; default is \code{c(-6, 6)}}
+#'     \item{theta_lim}{range of integration grid for each dimension; default is \code{c(-6, 6)}. Note that
+#'       when \code{itemtype = 'ULL'} a log-normal distribution is used and the range is change to
+#'       \code{c(.01, and 6^2)}, where the second term is the square of the \code{theta_lim} input instead}
 #'     \item{set.seed}{seed number used during estimation. Default is 12345}
 #'     \item{SEtol}{standard error tolerance criteria for the S-EM and MHRM computation of the
 #'       information matrix. Default is 1e-3}
@@ -583,6 +612,11 @@
 #'       positive or negative infinity. The default is \code{FALSE}}
 #'     \item{customTheta}{a custom \code{Theta} grid, in matrix form, used for integration.
 #'       If not defined, the grid is determined internally based on the number of \code{quadpts}}
+#'     \item{nconstrain}{same specification as the \code{constrain} list argument,
+#'       however imposes a negative equality constraint instead (e.g., \eqn{a12 = -a21}, which
+#'       is specified as \code{nconstrain = list(c(12, 21))}). Note that each specification
+#'       in the list must be of length 2, where the second element is taken to be -1 times the
+#'       first element}
 #'     \item{delta}{the deviation term used in numerical estimates when computing the ACOV matrix
 #'       with the 'forward' or 'central' numerical approaches, as well as Oakes' method with the
 #'       Richardson extrapolation. Default is 1e-5}
@@ -681,6 +715,10 @@
 #'
 #' Lord, F. M. & Novick, M. R. (1968). Statistical theory of mental test scores. Addison-Wesley.
 #'
+#' Lucke, J. F. (2015). Unipolar item response models. In S. P. Reise & D. A. Revicki
+#' (Eds.), Handbook of item response theory modeling: Applications to
+#' typical performance assessment (pp. 272-284). New York, NY:  Routledge/Taylor & Francis Group.
+#'
 #' Ramsay, J. O. (1975). Solving implicit equations in psychometric data analysis.
 #' \emph{Psychometrika, 40}, 337-360.
 #'
@@ -690,6 +728,10 @@
 #' Roberts, J. S., Donoghue, J. R., & Laughlin, J. E. (2000).
 #' A General Item Response Theory Model for Unfolding Unidimensional Polytomous Responses.
 #' \emph{Applied Psychological Measurement, 24}, 3-32.
+#'
+#' Shim, H., Bonifay, W., & Wiedermann, W. (2022). Parsimonious asymmetric item response
+#' theory modeling with the complementary log-log link. \emph{Behavior Research Methods, 55},
+#' 200-219.
 #'
 #' Maydeu-Olivares, A., Hernandez, A. & McDonald, R. P. (2006).
 #' A Multidimensional Ideal Point Item Response Theory Model for Binary Data.
@@ -737,7 +779,7 @@
 #' @export mirt
 #' @examples
 #'
-#' #load LSAT section 7 data and compute 1 and 2 factor models
+#' # load LSAT section 7 data and compute 1 and 2 factor models
 #' data <- expand.table(LSAT7)
 #' itemstats(data)
 #'
@@ -758,11 +800,11 @@
 #' plot(mod2, type = 'info') #test information
 #' plot(mod2, MI=200) #expected total score with 95% confidence intervals
 #'
-#' #estimated 3PL model for item 5 only
+#' # estimated 3PL model for item 5 only
 #' (mod1.3PL <- mirt(data, 1, itemtype = c('2PL', '2PL', '2PL', '2PL', '3PL')))
 #' coef(mod1.3PL)
 #'
-#' #internally g and u pars are stored as logits, so usually a good idea to include normal prior
+#' # internally g and u pars are stored as logits, so usually a good idea to include normal prior
 #' #  to help stabilize the parameters. For a value around .182 use a mean
 #' #  of -1.5 (since 1 / (1 + exp(-(-1.5))) == .182)
 #' model <- 'F = 1-5
@@ -772,12 +814,12 @@
 #' #limited information fit statistics
 #' M2(mod1.3PL.norm)
 #'
-#' #unidimensional ideal point model
+#' # unidimensional ideal point model
 #' idealpt <- mirt(data, 1, itemtype = 'ideal')
 #' plot(idealpt, type = 'trace', facet_items = TRUE)
 #' plot(idealpt, type = 'trace', facet_items = FALSE)
 #'
-#' #two factors (exploratory)
+#' # two factors (exploratory)
 #' mod2 <- mirt(data, 2)
 #' coef(mod2)
 #' summary(mod2, rotate = 'oblimin') #oblimin rotation
@@ -791,7 +833,7 @@
 #' scorestable <- fscores(mod2, full.scores = FALSE) #save factor score table
 #' head(scorestable)
 #'
-#' #confirmatory (as an example, model is not identified since you need 3 items per factor)
+#' # confirmatory (as an example, model is not identified since you need 3 items per factor)
 #' # Two ways to define a confirmatory model: with mirt.model, or with a string
 #'
 #' # these model definitions are equivalent
@@ -805,11 +847,11 @@
 #' # cmod <- mirt(data, cmodel2) # same as above
 #' coef(cmod)
 #' anova(cmod, mod2)
-#' #check if identified by computing information matrix
+#' # check if identified by computing information matrix
 #' (cmod <- mirt(data, cmodel, SE = TRUE))
 #'
 #' ###########
-#' #data from the 'ltm' package in numeric format
+#' # data from the 'ltm' package in numeric format
 #' itemstats(Science)
 #'
 #' pmod1 <- mirt(Science, 1)
@@ -818,8 +860,8 @@
 #' plot(pmod1, type = 'itemscore')
 #' summary(pmod1)
 #'
-#' #Constrain all slopes to be equal with the constrain = list() input or mirt.model() syntax
-#' #first obtain parameter index
+#' # Constrain all slopes to be equal with the constrain = list() input or mirt.model() syntax
+#' # first obtain parameter index
 #' values <- mirt(Science,1, pars = 'values')
 #' values #note that slopes are numbered 1,5,9,13, or index with values$parnum[values$name == 'a1']
 #' (pmod1_equalslopes <- mirt(Science, 1, constrain = list(c(1,5,9,13))))
@@ -840,30 +882,44 @@
 #' itemplot(pmod2, 1, rotate = 'oblimin')
 #' anova(pmod1, pmod2)
 #'
-#' #unidimensional fit with a generalized partial credit and nominal model
+#' # unidimensional fit with a generalized partial credit and nominal model
 #' (gpcmod <- mirt(Science, 1, 'gpcm'))
 #' coef(gpcmod)
 #'
-#' #for the nominal model the lowest and highest categories are assumed to be the
+#' # for the nominal model the lowest and highest categories are assumed to be the
 #' #  theoretically lowest and highest categories that related to the latent trait(s)
 #' (nomod <- mirt(Science, 1, 'nominal'))
 #' coef(nomod) #ordering of ak values suggest that the items are indeed ordinal
 #' anova(gpcmod, nomod)
 #' itemplot(nomod, 3)
 #'
-#' #generalized graded unfolding model
+#' # generalized graded unfolding model
 #' (ggum <- mirt(Science, 1, 'ggum'))
 #' coef(ggum, simplify=TRUE)
 #' plot(ggum)
 #' plot(ggum, type = 'trace')
 #' plot(ggum, type = 'itemscore')
 #'
-#' #monotonic polyomial models
+#' # monotonic polyomial models
 #' (monopoly <- mirt(Science, 1, 'monopoly'))
 #' coef(monopoly, simplify=TRUE)
 #' plot(monopoly)
 #' plot(monopoly, type = 'trace')
 #' plot(monopoly, type = 'itemscore')
+#'
+#' # unipolar IRT model
+#' unimod <- mirt(Science, itemtype = 'ULL')
+#' coef(unimod, simplify=TRUE)
+#' plot(unimod)
+#' plot(unimod, type = 'trace')
+#' itemplot(unimod, 1)
+#'
+#' # following use the correct log-normal density for latent trait
+#' itemfit(unimod)
+#' M2(unimod, type = 'C2')
+#' fs <- fscores(unimod)
+#' hist(fs, 20)
+#' fscores(unimod, method = 'EAPsum', full.scores = FALSE)
 #'
 #' ## example applying survey weights.
 #' # weight the first half of the cases to be more representative of population
@@ -873,7 +929,7 @@
 #' weighted <- mirt(Science, 1, survey.weights=survey.weights)
 #'
 #' ###########
-#' #empirical dimensionality testing that includes 'guessing'
+#' # empirical dimensionality testing that includes 'guessing'
 #'
 #' data(SAT12)
 #' data <- key2binary(SAT12,
@@ -883,28 +939,28 @@
 #' mod1 <- mirt(data, 1)
 #' extract.mirt(mod1, 'time') #time elapsed for each estimation component
 #'
-#' #optionally use Newton-Raphson for (generally) faster convergence in the M-step's
+#' # optionally use Newton-Raphson for (generally) faster convergence in the M-step's
 #' mod1 <- mirt(data, 1, optimizer = 'NR')
 #' extract.mirt(mod1, 'time')
 #'
 #' mod2 <- mirt(data, 2, optimizer = 'NR')
-#' #difficulty converging with reduced quadpts, reduce TOL
+#' # difficulty converging with reduced quadpts, reduce TOL
 #' mod3 <- mirt(data, 3, TOL = .001, optimizer = 'NR')
 #' anova(mod1,mod2)
 #' anova(mod2, mod3) #negative AIC, 2 factors probably best
 #'
-#' #same as above, but using the QMCEM method for generally better accuracy in mod3
+#' # same as above, but using the QMCEM method for generally better accuracy in mod3
 #' mod3 <- mirt(data, 3, method = 'QMCEM', TOL = .001, optimizer = 'NR')
 #' anova(mod2, mod3)
 #'
-#' #with fixed guessing parameters
+#' # with fixed guessing parameters
 #' mod1g <- mirt(data, 1, guess = .1)
 #' coef(mod1g)
 #'
 #' ###########
-#' #graded rating scale example
+#' # graded rating scale example
 #'
-#' #make some data
+#' # make some data
 #' set.seed(1234)
 #' a <- matrix(rep(1, 10))
 #' d <- matrix(c(1,0.5,-.5,-1), 10, 4, byrow = TRUE)
@@ -926,12 +982,12 @@
 #' SAT12[SAT12 == 8] <- NA #set 8 as a missing value
 #' head(SAT12)
 #'
-#' #correct answer key
+#' # correct answer key
 #' key <- c(1,4,5,2,3,1,2,1,3,1,2,4,2,1,5,3,4,4,1,4,3,3,4,1,3,5,1,3,1,5,4,5)
 #' scoredSAT12 <- key2binary(SAT12, key)
 #' mod0 <- mirt(scoredSAT12, 1)
 #'
-#' #for first 5 items use 2PLNRM and nominal
+#' # for first 5 items use 2PLNRM and nominal
 #' scoredSAT12[,1:5] <- as.matrix(SAT12[,1:5])
 #' mod1 <- mirt(scoredSAT12, 1, c(rep('nominal',5),rep('2PL', 27)))
 #' mod2 <- mirt(scoredSAT12, 1, c(rep('2PLNRM',5),rep('2PL', 27)), key=key)
@@ -942,7 +998,7 @@
 #' itemplot(mod1, 1)
 #' itemplot(mod2, 1)
 #'
-#' #compare added information from distractors
+#' # compare added information from distractors
 #' Theta <- matrix(seq(-4,4,.01))
 #' par(mfrow = c(2,3))
 #' for(i in 1:5){
@@ -953,7 +1009,7 @@
 #' }
 #' par(mfrow = c(1,1))
 #'
-#' #test information
+#' # test information
 #' plot(Theta, testinfo(mod2, Theta), type = 'l', main = 'Test information', ylab = 'Information')
 #' lines(Theta, testinfo(mod0, Theta), col = 'red')
 #'
@@ -963,9 +1019,9 @@
 #' fulldata <- expand.table(LSAT7)
 #' (mod1 <- mirt(fulldata, 1, method = 'MHRM'))
 #'
-#' #Confirmatory models
+#' # Confirmatory models
 #'
-#' #simulate data
+#' # simulate data
 #' a <- matrix(c(
 #' 1.5,NA,
 #' 0.5,NA,
@@ -991,15 +1047,15 @@
 #' items <- c(rep('2PL',4), rep('graded',3), '2PL')
 #' dataset <- simdata(a,d,2000,items,sigma)
 #'
-#' #analyses
-#' #CIFA for 2 factor crossed structure
+#' # analyses
+#' # CIFA for 2 factor crossed structure
 #'
 #' model.1 <- '
 #'   F1 = 1-4
 #'   F2 = 4-8
 #'   COV = F1*F2'
 #'
-#' #compute model, and use parallel computation of the log-likelihood
+#' # compute model, and use parallel computation of the log-likelihood
 #' if(interactive()) mirtCluster()
 #' mod1 <- mirt(dataset, model.1, method = 'MHRM')
 #' coef(mod1)
@@ -1007,7 +1063,7 @@
 #' residuals(mod1)
 #'
 #' #####
-#' #bifactor
+#' # bifactor
 #' model.3 <- '
 #'   G = 1-8
 #'   F1 = 1-4
@@ -1020,7 +1076,7 @@
 #' anova(mod1,mod3)
 #'
 #' #####
-#' #polynomial/combinations
+#' # polynomial/combinations
 #' data(SAT12)
 #' data <- key2binary(SAT12,
 #'                   key = c(1,4,5,2,3,1,2,1,3,1,2,4,2,1,5,3,4,4,1,4,3,3,4,1,3,5,1,3,1,5,4,5))
@@ -1040,7 +1096,7 @@
 #' (mod.combo <- mirt(data, model.combo))
 #' anova(mod.combo, mod.quad)
 #'
-#' #non-linear item and test plots
+#' # non-linear item and test plots
 #' plot(mod.quad)
 #' plot(mod.combo, type = 'SE')
 #' itemplot(mod.quad, 1, type = 'score')
@@ -1048,7 +1104,7 @@
 #' itemplot(mod.combo, 2, type = 'infocontour')
 #'
 #' ## empirical histogram examples (normal, skew and bimodality)
-#' #make some data
+#' # make some data
 #' set.seed(1234)
 #' a <- matrix(rlnorm(50, .2, .2))
 #' d <- matrix(rnorm(50))
@@ -1078,16 +1134,16 @@
 #' dat <- expand.table(LSAT6)
 #' itemstats(dat)
 #'
-#' #free latent mean and variance terms
+#' # free latent mean and variance terms
 #' model <- 'Theta = 1-5
 #'           MEAN = Theta
 #'           COV = Theta*Theta'
 #'
-#' #view how vector of parameters is organized internally
+#' # view how vector of parameters is organized internally
 #' sv <- mirt(dat, model, itemtype = 'Rasch', pars = 'values')
 #' sv[sv$est, ]
 #'
-#' #constraint: create function for solnp to compute constraint, and declare value in eqB
+#' # constraint: create function for solnp to compute constraint, and declare value in eqB
 #' eqfun <- function(p, optim_args) sum(p[1:5]) #could use browser() here, if it helps
 #' LB <- c(rep(-15, 6), 1e-4) # more reasonable lower bound for variance term
 #'
@@ -1104,7 +1160,7 @@
 #' #######
 #' # latent regression Rasch model
 #'
-#' #simulate data
+#' # simulate data
 #' set.seed(1234)
 #' N <- 1000
 #'
@@ -1113,26 +1169,26 @@
 #' covdata <- data.frame(X1, X2)
 #' Theta <- matrix(0.5 * X1 + -1 * X2 + rnorm(N, sd = 0.5))
 #'
-#' #items and response data
+#' # items and response data
 #' a <- matrix(1, 20); d <- matrix(rnorm(20))
 #' dat <- simdata(a, d, 1000, itemtype = '2PL', Theta=Theta)
 #'
-#' #unconditional Rasch model
+#' # unconditional Rasch model
 #' mod0 <- mirt(dat, 1, 'Rasch')
 #'
-#' #conditional model using X1 and X2 as predictors of Theta
+#' # conditional model using X1 and X2 as predictors of Theta
 #' mod1 <- mirt(dat, 1, 'Rasch', covdata=covdata, formula = ~ X1 + X2)
 #' coef(mod1, simplify=TRUE)
 #' anova(mod0, mod1)
 #'
-#' #bootstrapped confidence intervals
+#' # bootstrapped confidence intervals
 #' boot.mirt(mod1, R=5)
 #'
-#' #draw plausible values for secondary analyses
+#' # draw plausible values for secondary analyses
 #' pv <- fscores(mod1, plausible.draws = 10)
 #' pvmods <- lapply(pv, function(x, covdata) lm(x ~ covdata$X1 + covdata$X2),
 #'                  covdata=covdata)
-#' #population characteristics recovered well, and can be averaged over
+#' # population characteristics recovered well, and can be averaged over
 #' so <- lapply(pvmods, summary)
 #' so
 #'
@@ -1185,6 +1241,31 @@
 #'
 #' itemfit(dav) # assume normal prior
 #' itemfit(dav, use_dentype_estimate=TRUE) # use Davidian estimated prior shape
+#'
+#' ###########
+#' # 5PL and restricted 5PL example
+#' dat <- expand.table(LSAT7)
+#'
+#' mod2PL <- mirt(dat)
+#' mod2PL
+#'
+#' # Following does not converge without including strong priors
+#' # mod5PL <- mirt(dat, itemtype = '5PL')
+#' # mod5PL
+#'
+#' # restricted version of 5PL (asymmetric 2PL)
+#' model <- 'Theta = 1-5
+#'           FIXED = (1-5, g), (1-5, u)'
+#'
+#' mod2PL_asym <- mirt(dat, model=model, itemtype = '5PL')
+#' mod2PL_asym
+#' coef(mod2PL_asym, simplify=TRUE)
+#' coef(mod2PL_asym, simplify=TRUE, IRTpars=TRUE)
+#'
+#' # no big difference statistically or visually
+#' anova(mod2PL, mod2PL_asym)
+#' plot(mod2PL, type = 'trace')
+#' plot(mod2PL_asym, type = 'trace')
 #'
 #' }
 mirt <- function(data, model = 1, itemtype = NULL, guess = 0, upper = 1, SE = FALSE,